+++ thunberg-wine/src/App.css
@@ -12,33 +12,3 @@
   border-radius: 10px;
 }
 
-.ProductCard {
-  display: flex;
-  flex-direction: column; 
-  align-items: center; 
-  border: 1px solid #ddd;
-  border-radius: 4px;
-  padding: 16px;
-  margin: 10px;
-  text-align: center;
-}
-
-.ProductCard img {
-  width: 100%; 
-  max-width: 220px; 
-  height: auto; 
-}
-
-.products-grid{
-
-  display: flex;
-  justify-content: space-around;
-  
-}
-
-.products-grid img{
-  border: 2px solid black;
-  border-radius: 5px;
-} 
-
-
