+++ thunberg-wine/src/App.tsx
@@ -1,9 +1,10 @@
 import './App.css';
 import { BrowserRouter } from "react-router-dom";
 import MainRoutes from "./Routes/mainRoutes";
-import RoundedButton from './Components/RoundedBotton';
+import RoundedButton from './Components/RoundedButton/RoundedBotton';
 import { CartProvider } from './context/CartContext';
 import { Header } from './Components/Header/Header';
+import { Nav1 } from './Components/Nav/Nav';
 
 function App() {
   // return <GridCard />
