+++ thunberg-wine/src/Components/Header/Header.css
@@ -1,4 +1,3 @@
-.right-buttons{
-
-    padding: 20%;
+header {
+    width: 100%;
 }
