+++ thunberg-wine/src/Components/Header/Header.tsx
@@ -1,28 +1,12 @@
-import { useNavigate } from "react-router-dom";
-import { useCart } from "../../context/CartContext";
-import RoundedButton from "../RoundedBotton";
-import CartIcon from "../../assets/shopping-cart.svg"
+
 import '../../Components/Header/Header.css';
+import { Nav1 } from "../Nav/Nav";
 
 export function Header() {
-    const {totalItems} = useCart();
-    const navigate = useNavigate();
+   
     return(
         <header className="header">
-            
-            <div className="right-buttons">
-
-                <RoundedButton 
-                label2="Aggiunti al carrello"
-                label= {totalItems>0? totalItems.toString() : "0"}
-                icon={CartIcon}
-                onclick={() => {
-                    navigate("/cart");
-                }}
-                />
-
-
-            </div>
+            <Nav1/>
         </header>
     );
 }
