+++ thunberg-wine/src/Components/ProductsCard/GiftCard.css
@@ -1,0 +1,6 @@
+.products-flex {
+    width: 100%;
+    display: flex;
+    justify-content: space-around;
+    flex-wrap: wrap;
+}
