+++ thunberg-wine/src/Components/RoundedButton/RoundedButton.css
@@ -1,0 +1,8 @@
+.button-boh {
+    padding: 0;
+    display: flex;
+    align-items: center;
+    justify-content: space-around;
+    background-color: rgb(106, 6, 6);
+    width: 7%;
+}
