+++ thunberg-wine/src/Interfaces/productInterfaces.ts
@@ -13,31 +13,91 @@ export const mockProducts: Product[] = [
     {
         id: "1",
         name: "Product 1",
-        image: require ("../assets/img/Vino-rosso.png"),
+        image: require ("../assets/img/vino-rosato.png"),
         price: "9,99€",
         description: "the wine ever",
-        width: "220",
-        height: "230",
+        width: "200",
+        height: "200",
     },
 
     {
         id: "2",
         name: "Product 2",
-        image: require ("../assets/img/vino-bianco.jpg"),
+        image: require ("../assets/img/vino-bianco.webp"),
         price: "9,99€",
         description: "the wine ever",
-        width: "180",
-        height: "230",
+        width: "200",
+        height: "200",
     },
 
     {
         id: "3",
         name: "Product 4",
-        image: require ("../assets/img/vino-rosato.jpg"),
+        image: require ("../assets/img/vino-rosso.png"),
         price: "9,99€",
         description: "the wine ever",
         width: "200",
-        height: "230",
-    }
+        height: "200",
+    },
+
+    {
+        id: "1",
+        name: "Product 1",
+        image: require ("../assets/img/vino-rosato.png"),
+        price: "9,99€",
+        description: "the wine ever",
+        width: "200",
+        height: "200",
+    },
+
+    {
+        id: "2",
+        name: "Product 2",
+        image: require ("../assets/img/vino-bianco.webp"),
+        price: "9,99€",
+        description: "the wine ever",
+        width: "200",
+        height: "200",
+    },
+
+    {
+        id: "3",
+        name: "Product 4",
+        image: require ("../assets/img/vino-rosso.png"),
+        price: "9,99€",
+        description: "the wine ever",
+        width: "200",
+        height: "200",
+    },
+
+    {
+        id: "1",
+        name: "Product 1",
+        image: require ("../assets/img/vino-rosato.png"),
+        price: "9,99€",
+        description: "the wine ever",
+        width: "200",
+        height: "200",
+    },
+
+    {
+        id: "2",
+        name: "Product 2",
+        image: require ("../assets/img/vino-bianco.webp"),
+        price: "9,99€",
+        description: "the wine ever",
+        width: "200",
+        height: "200",
+    },
+
+    {
+        id: "3",
+        name: "Product 4",
+        image: require ("../assets/img/vino-rosso.png"),
+        price: "9,99€",
+        description: "the wine ever",
+        width: "200",
+        height: "200",
+    },
 
 ];
